% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygeo.R
\name{tidygeo}
\alias{tidygeo}
\title{Extract a tidy data frame from geoRSS, geo-Atom and geoJSON feeds}
\usage{
tidygeo(feed, config = list(), clean_tags = TRUE, list = FALSE)
}
\arguments{
\item{clean_tags}{default \code{TRUE}. Clean HTML tags from summary and content?}

\item{list}{default \code{FALSE}. Return metadata and entry information in separate dataframes?
These will be combined in a named list.}
}
\description{
Extract a tidy data frame from geoRSS, geo-Atom and geoJSON feeds
}
